--- src/main/java/com/bit/joe/shoppingmall/security/handler/CustomLogoutHandler.java
@@ -1,18 +1,0 @@
-package com.bit.joe.shoppingmall.security.handler;
-
-import org.springframework.security.core.Authentication;
-import org.springframework.security.web.authentication.logout.LogoutHandler;
-import org.springframework.stereotype.Component;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-@Component
-public class CustomLogoutHandler implements LogoutHandler {
-
-    @Override
-    public void logout(
-            HttpServletRequest request,
-            HttpServletResponse response,
-            Authentication authentication) {}
-}
