+++ src/main/java/com/bit/joe/shoppingmall/view/AdminController.java
@@ -1,14 +1,14 @@
-package com.bit.joe.shoppingmall.controller;
+package com.bit.joe.shoppingmall.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/history")
-public class HistoryController {
+@RequestMapping("/admin")
+public class AdminController {
     @GetMapping
-    public String history() {
-        return "product/history";
+    public String index() {
+        return "admin/insight";
     }
 }
