+++ src/main/java/com/bit/joe/shoppingmall/view/CartHtmlController.java
@@ -1,14 +1,14 @@
-package com.bit.joe.shoppingmall.controller;
+package com.bit.joe.shoppingmall.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/admin")
-public class AdminController {
+@RequestMapping("/cart")
+public class CartHtmlController {
     @GetMapping
-    public String index() {
-        return "admin/insight";
+    public String cart() {
+        return "product/cart";
     }
 }
