+++ src/main/java/com/bit/joe/shoppingmall/view/HistoryController.java
@@ -1,14 +1,14 @@
-package com.bit.joe.shoppingmall.controller;
+package com.bit.joe.shoppingmall.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/product")
-public class ProductdetailController {
+@RequestMapping("/history")
+public class HistoryController {
     @GetMapping
-    public String productdetail() {
-        return "product/product-detail";
+    public String history() {
+        return "product/history";
     }
 }
