+++ src/main/java/com/bit/joe/shoppingmall/view/HomeController.java
@@ -1,4 +1,4 @@
-package com.bit.joe.shoppingmall.controller;
+package com.bit.joe.shoppingmall.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
