+++ src/main/java/com/bit/joe/shoppingmall/view/MemberController.java
@@ -1,14 +1,14 @@
-package com.bit.joe.shoppingmall.controller;
+package com.bit.joe.shoppingmall.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/admin/inventory")
-public class InventoryController {
+@RequestMapping("/admin/member")
+public class MemberController {
     @GetMapping
     public String index() {
-        return "admin/inventory";
+        return "admin/member";
     }
 }
