+++ src/main/java/com/bit/joe/shoppingmall/view/ProductdetailController.java
@@ -1,14 +1,14 @@
-package com.bit.joe.shoppingmall.controller;
+package com.bit.joe.shoppingmall.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/cart")
-public class CartHtmlController {
+@RequestMapping("/product")
+public class ProductdetailController {
     @GetMapping
-    public String cart() {
-        return "product/cart";
+    public String productdetail() {
+        return "product/product-detail";
     }
 }
