+++ src/main/resources/application.properties
@@ -18,7 +18,6 @@ spring.jpa.hibernate.ddl-auto=update
 
 spring.devtools.livereload.enabled=true
 
-spring.thymeleaf.prefix=classpath:/templates/thymeleaf/
 spring.thymeleaf.suffix=.html
 spring.thymeleaf.mode=HTML
 spring.thymeleaf.cache=false
