+++ src/main/resources/templates/admin/insight.html
