+++ src/main/resources/templates/admin/inventory.html
