+++ src/main/resources/templates/common/footer.html
