+++ src/main/resources/templates/common/navigation.html
