+++ src/main/resources/templates/index.html
@@ -14,7 +14,7 @@
     <ul></ul>
 
     <script>
-      fetch("http://localhost:8080/product/get-all")
+      fetch("/product/get-all")
         .then((res) => res.json())
         .then((data) => {
           console.log(data);
