+++ src/main/resources/templates/product/product-detail.html
