--- src/main/resources/templates/thymeleaf/user/login.html
@@ -1,22 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0"
-    />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <title>Document</title>
-  </head>
-  <body>
-    <h1>로그인</h1>
-    <hr />
-    <p>이메일 : <input type="text" /></p>
-    <p>비밀번호 : <input type="text" /></p>
-    <br />
-    <input type="submit" value="로그인하기" />
-    <br />
-    <p><a href="/user/signUp">아직 회원이 아니신가요?</a></p>
-  </body>
-</html>
