+++ src/main/resources/templates/user/mainPage.html
@@ -1,0 +1,39 @@
+<!doctype html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org">
+<head>
+  <meta charset="UTF-8"/>
+  <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+  <title>MainPage</title>
+  <link
+          href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css"
+          rel="stylesheet"
+  />
+</head>
+<body>
+<!-- Corrected th:replace to properly include the navbar fragment -->
+<div th:replace="~{common/navigation :: navbar}"></div>
+
+<hr/>
+<div>
+  <label>
+    <input type="text"/>
+  </label>
+  <span><input type="submit" value="입력"/></span>
+</div>
+<div id="mainImage">
+  <h1>이미지(이벤트로 넘기기)</h1>
+</div>
+<br/>
+<div class="container">
+  <div><img src="" alt="이미지1"/></div>
+  <div><img src="" alt="이미지2"/></div>
+  <div><img src="" alt="이미지3"/></div>
+  <div><img src="" alt="이미지4"/></div>
+  <div><img src="" alt="이미지5"/></div>
+  <div><img src="" alt="이미지6"/></div>
+</div>
+
+<div th:replace="~{common/footer :: footer}"></div>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"></script>
+</body>
+</html>
