+++ src/test/resources/junit-platform.properties
@@ -1,5 +1,5 @@
 junit.jupiter.execution.parallel.enabled=true
-junit.jupiter.execution.parallel.mode.classes.default==concurrent
+junit.jupiter.execution.parallel.mode.classes.default=concurrent
 junit.jupiter.execution.parallel.mode.default=same_thread
 junit.jupiter.execution.parallel.config.strategy=dynamic
 junit.jupiter.execution.parallel.config.dynamic.factor=1
